requirements.txt
PyYaml>=3.12
Docker>=3.0.0
jsonschema>=2.6.0 
